# backend/docs/swagger.yaml
openapi: 3.0.1
info:
  description: "This is documentation for Kalorické tabulky, describing its REST API"
  version: 1.0.0
  title: "Kalorické tabulky"
  contact:
    email: "514017@mail.muni.cz"
servers:
  - url: "http://localhost:3000/"
...
paths:
...
  /api/food/:{id}:
    delete:
      tags:
        - "Food"
      summary: "Delete food by id by setting deleted parabeter to true"
      operationId: "deleteFood"
      parameters:
        - in: path
          name: id  
          required: true
          schema:
            type: uuid
          description: The user UUID
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SingleFoundResponse"
        "500":
          description: "internal error"
...
components:
  schemas:
...
    SingleFoundResponse:
      type: object
      properties: 
        status:
          type: string
        data:
          type: FoodData
        message:
          type: string
...